Revert "Message center"
