+++ android/app/build.gradle
@@ -197,7 +197,7 @@ def getKeyStorePassword() {
 
 android {
     compileSdkVersion 23
-    buildToolsVersion '26.0.2'
+    buildToolsVersion '26.0.0'
     project.archivesBaseName = appName;
 
     lintOptions {
