+++ app/api/constants/user.js
@@ -1,29 +1,7 @@
 export default {
   email: 'usman.khalid@digi.com.my',
   password: '123',
-  chatbotImage: 'https://i.imgur.com/7tvqC6m.png',
-  userImage: 'https://i.imgur.com/bQCTmCH.jpg',
-  userName: 'Roshini Vellu',
-  userTitle: 'Senior Associate',
   helper_consts: {
-    ACTION_CENTER: {
-      OK: 1,
-      CONNETION_TIMEDOUT: 2,
-      RESOURCE_NOT_FOUND: 3,
-      SERVER_PROBLEM: 4,
-    },
-    MESSAGE_CENTER: {
-      OK: 1,
-      CONNETION_TIMEDOUT: 2,
-      RESOURCE_NOT_FOUND: 3,
-      SERVER_PROBLEM: 4,
-    },
-    DASHBOARD: {
-      OK: 1,
-      CONNETION_TIMEDOUT: 2,
-      RESOURCE_NOT_FOUND: 3,
-      SERVER_PROBLEM: 4,
-    },
     RESET_PASSWORD: {
       OK: 1,
       CONNETION_TIMEDOUT: 2,
