--- app/api/user/data/action_center_data.js
@@ -1,102 +1,0 @@
-export default (data = [
-  {
-    name: 'Ali Haider',
-    request: {
-      description: 'Accomodation (Overseas)',
-      detail: 'Total: MYR 100.00',
-      status: 'Review',
-      image: 'https://i.imgur.com/6DkmJ9t.jpg',
-    },
-  },
-  {
-    name: 'Haroon Bhatti',
-    request: {
-      description: 'Meals - Breakfast (Local)',
-      detail: 'MYR 15.00',
-      status: 'Approved',
-      image: 'https://i.imgur.com/fU4XyXo.jpg',
-    },
-  },
-  {
-    name: 'Dira',
-    request: {
-      description: 'Holidays are approaching!',
-      detail:
-        'Hey there! I noticed that Christmas is coming and Air Asia is having a promotion on certain flights. Do you want me to help to book a flight?',
-      status: '',
-      image: 'https://i.imgur.com/7tvqC6m.png',
-    },
-  },
-  {
-    name: 'Haroon Bhatti',
-    request: {
-      description: 'Entertainment Biz Associate',
-      detail: 'MYR 1500.00',
-      status: 'Rejected',
-      image: 'https://i.imgur.com/fU4XyXo.jpg',
-    },
-  },
-  {
-    name: 'Roshini',
-    request: {
-      description: 'Gadget Fund',
-      detail: 'MYR 150.00',
-      status: 'Rejected',
-      image: 'https://i.imgur.com/bQCTmCH.jpg',
-    },
-  },
-  {
-    name: 'Haroon Bhatti',
-    request: {
-      description: 'Entertainment Biz Associate',
-      detail: 'MYR 1500.00',
-      status: 'Rejected',
-      image: 'https://i.imgur.com/fU4XyXo.jpg',
-    },
-  },
-  {
-    name: 'Roshini',
-    request: {
-      description: 'Gadget Fund',
-      detail: 'MYR 150.00',
-      status: 'Rejected',
-      image: 'https://i.imgur.com/bQCTmCH.jpg',
-    },
-  },
-  {
-    name: 'Haroon Bhatti',
-    request: {
-      description: 'Entertainment Biz Associate',
-      detail: 'MYR 1500.00',
-      status: 'Rejected',
-      image: 'https://i.imgur.com/fU4XyXo.jpg',
-    },
-  },
-  {
-    name: 'Roshini',
-    request: {
-      description: 'Gadget Fund',
-      detail: 'MYR 150.00',
-      status: 'Rejected',
-      image: 'https://i.imgur.com/bQCTmCH.jpg',
-    },
-  },
-  {
-    name: 'Haroon Bhatti',
-    request: {
-      description: 'Entertainment Biz Associate',
-      detail: 'MYR 1500.00',
-      status: 'Rejected',
-      image: 'https://i.imgur.com/fU4XyXo.jpg',
-    },
-  },
-  {
-    name: 'Roshini',
-    request: {
-      description: 'Gadget Fund',
-      detail: 'MYR 150.00',
-      status: 'Rejected',
-      image: 'https://i.imgur.com/bQCTmCH.jpg',
-    },
-  },
-]);
