--- app/api/user/data/message_center_data.js
@@ -1,52 +1,0 @@
-export default (data = [
-  {
-    title: 'Welcome to alt.hr',
-    description:
-      'Congratulations! You have completed your onboarding journey. You can now access all modules of alt.hr such as Expenses and Travels. Congratulations! You have completed your onboarding journey. You can now access all modules of alt.hr such as Expenses and Travels. Congratulations! You have completed your onboarding journey. You can now access all modules of alt.hr such as Expenses and Travels. Congratulations! You have completed your onboarding journey. You can now access all modules of alt.hr such as Expenses and Travels. Congratulations! You have completed your onboarding journey. You can now access all modules of alt.hr such as Expenses and Travels. Congratulations! You have completed your onboarding journey. You can now access all modules of alt.hr such as Expenses and Travels. Congratulations! You have completed your onboarding journey. You can now access all modules of alt.hr such as Expenses and Travels. Congratulations! You have completed your onboarding journey. You can now access all modules of alt.hr such as Expenses and Travels. Congratulations! You have completed your onboarding journey. You can now access all modules of alt.hr such as Expenses and Travels. Congratulations! You have completed your onboarding journey. You can now access all modules of alt.hr such as Expenses and Travels.',
-    date: '27 Mar 2018',
-    imageUrl: '',
-    videoUrl: '',
-  },
-  {
-    title: 'Building an Innovation 360 culture',
-    description:
-      'Hey Digizens. At Townhall, Albern shared how innovation belongs to everyone and each of us has a role to play in building a culture of innovation.',
-    date: '26 Mar 2018',
-    imageUrl:
-      'https://www.gettyimages.ie/gi-resources/images/Homepage/Hero/UK/CMS_Creative_164657191_Kingfisher.jpg',
-    videoUrl: '',
-  },
-  {
-    title: 'Message from the CEO',
-    description:
-      'The CEO of Digi, Albern Murty, welcomes you to be part of the Digi family and shares his valuable experience and advice for all the newbies coming onboard',
-    date: '20 Mar 2018',
-    imageUrl: '',
-    videoUrl: 'https://www.youtube.com/embed/vyM5_YpYJH0',
-  },
-  {
-    title: 'Welcome to alt.hr',
-    description:
-      'Congratulations! You have completed your onboarding journey. You can now access all modules of alt.hr such as Expenses and Travels.',
-    date: '27 Mar 2018',
-    imageUrl: '',
-    videoUrl: '',
-  },
-  {
-    title: 'Building an Innovation 360 culture',
-    description:
-      'Hey Digizens. At Townhall, Albern shared how innovation belongs to everyone and each of us has a role to play in building a culture of innovation.',
-    date: '26 Mar 2018',
-    imageUrl:
-      'https://www.gettyimages.ie/gi-resources/images/Homepage/Hero/UK/CMS_Creative_164657191_Kingfisher.jpg',
-    videoUrl: '',
-  },
-  {
-    title: 'Message from the CEO',
-    description:
-      'The CEO of Digi, Albern Murty, welcomes you to be part of the Digi family and shares his valuable experience and advice for all the newbies coming onboard',
-    date: '20 Mar 2018',
-    imageUrl: '',
-    videoUrl: 'https://www.youtube.com/embed/vyM5_YpYJH0',
-  },
-]);
