+++ app/api/user/index.js
@@ -1,14 +1,8 @@
 import userConstants from '../constants/user';
 import AuthInfo from '@digihr_models/user/auth_info';
 import _ from 'lodash';
-import { getActionCenter } from './get_action_center_data';
-import { getMessageCenter } from './get_message_center_data';
-import { getDashboard } from './get_dashboard_data';
 import { performResetPassword } from './reset_password';
 import * as ResetPasswordProcessor from './processor/reset_password';
-import * as ActionCenterProcessor from './processor/action_center';
-import * as MessageCenterProcessor from './processor/message_center';
-import * as DashboardProcessor from './processor/dashboard';
 
 /**
  * Performs an authenticate user API call
@@ -44,42 +38,6 @@ export function resetPassword(
 ) {
   return performResetPassword(email, strategy).catch(e => {
     throw ResetPasswordProcessor.processError(e, email);
-  });
-}
-
-/**
- * Performs a get dashboard action center data API call
- * @param {userConstants.helper_consts.ACTION_CENTER} strategy
- */
-export function getActionCenterData(
-  strategy = userConstants.helper_consts.ACTION_CENTER.OK
-) {
-  return getActionCenter(strategy).catch(e => {
-    throw ActionCenterProcessor.processError(e);
-  });
-}
-
-/**
- * Performs a get dashboard message center data API call
- * @param {userConstants.helper_consts.MESSAGE_CENTER} strategy
- */
-export function getMessageCenterData(
-  strategy = userConstants.helper_consts.MESSAGE_CENTER.OK
-) {
-  return getMessageCenter(strategy).catch(e => {
-    throw MessageCenterProcessor.processError(e);
-  });
-}
-
-/**
- * Performs a get dashboard message center data API call
- * @param {userConstants.helper_consts.MESSAGE_CENTER} strategy
- */
-export function getDashboardData(
-  strategy = userConstants.helper_consts.MESSAGE_CENTER.OK
-) {
-  return getDashboard(strategy).catch(e => {
-    throw DashboardProcessor.processError(e);
   });
 }
