+++ app/assets/images.js
@@ -1,5 +1,6 @@
 const images = {
-  altLogo: require('./img/alt-logo.png'),
+	altLogo: require('./img/alt-logo.png'),
 };
+
 
 export default images;
