--- app/components/package.json
@@ -1,3 +1,0 @@
-{
-    "name": "@digihr_app_components"
-}
