+++ app/config/routes.js
@@ -6,8 +6,6 @@ import Login from '@digihr_pages/login';
 import Welcome from '@digihr_pages/welcome';
 import ResetPassword from '@digihr_pages/resetPassword';
 import Signup from '@digihr_pages/signup';
-import MessageCenter from '@digihr_pages/messageCenter';
-import MessageDetail from '@digihr_pages/messageDetail';
 import SignupVerification from '@digihr_pages/signupVerification';
 
 RouteManager.Instance.add(
@@ -15,12 +13,6 @@ RouteManager.Instance.add(
 );
 
 RouteManager.Instance.add(new Route('Dashboard', Dashboard, 'DashboardScreen'));
-RouteManager.Instance.add(
-  new Route('MessageCenter', MessageCenter, 'MessageCenterScreen')
-);
-RouteManager.Instance.add(
-  new Route('MessageDetail', MessageDetail, 'MessageDetail')
-);
 RouteManager.Instance.add(new Route('Login', Login, 'LoginScreen'));
 RouteManager.Instance.add(new Route('Welcome', Welcome, 'WelcomeScreen'));
 RouteManager.Instance.add(
