+++ app/index.js
@@ -1,6 +1,6 @@
 import { createReduxBoundAddListener } from 'react-navigation-redux-helpers';
 import { BackHandler } from 'react-native';
-import { addNavigationHelpers, NavigationActions } from 'react-navigation';
+import { addNavigationHelpers } from 'react-navigation';
 import { Provider, connect } from 'react-redux';
 import React, { Component } from 'react';
 import Store from '@digihr_lib/store';
