--- app/pages/messageCenter/view_controller.js
@@ -1,6 +1,0 @@
-import RouteConfig from '@digihr_app_config/routes';
-import { getMessageCenterData } from '@digihr_api/user';
-
-export function getMessageCenter() {
-  return getMessageCenterData(1);
-}
