+++ app/translations/en.js
@@ -1,49 +1,31 @@
 export default {
   about_us: 'About this app',
   accept: 'Accept',
-  action_center: 'ACTION CENTER',
-  alphanumeric_password: 'ALPHANUMERIC WITH MINIMUM 8 CHARACTERS',
   app_name: 'alt hr app',
   app_version: 'Version',
-  confirm_password: 'CONFIRM PASSWORD',
-  confirm_password_not_matched: 'Confirm password is not matched with password',
-  continue: 'CONTINUE',
-  dashboard: 'DASHBOARD',
   dont_have_an_account: "Don't have an account? ",
   email_address: 'Email address',
   email_not_valid: 'email is not valid',
   email_not_registered:
     'email is not registered under our system to change password',
   email_sent: 'Email has sent',
-  enable_biometric: 'ENABLE BIOMETRIC VERIFICATION',
-  enter_ic_passport: 'ENTER YOUR IC OR PASSPORT NUMBER',
-  enter_invitation_code: 'Please enter your invitation code',
   finish: 'Finish',
-  help_message: 'Hi, How may I help you today?',
   forgot_your_password: 'Forgot your password?',
-  ic_passport_number: 'IC/PASSPORT NUMBER',
   i_agree: 'I Agree',
   in_progress: 'Available soon',
   incorrect_email_password: 'Incorrect email or password',
   internet_error: 'Oh no! Check your connection and try again',
   invalid_email: 'email is not valid',
-  invitation_code: 'INVITATION CODE',
   license_agreement: 'License Agreement',
   log_off: 'Log off',
   logging_out_message: 'Logging you out ...',
   logging_you_in: 'Loggin you in ...',
   login: 'Login',
   logout_message: 'Logging out..',
-  message_center: 'MESSAGE CENTRE',
   no_internet: 'check your internet connection',
-  nothing_new_to_show: 'Nothing new to show right now...',
   password: 'Password',
-  password_not_meet_requirement:
-    'Password should contain at least one special character and one uppercase letter',
   password_reset: 'PASSWORD RESET',
-  policies: 'POLICIES',
   problem_with_server: 'Some problem with our server',
-  reenter_password: 'RE-ENTER PASSWORD',
   reset_password: 'RESET PASSWORD',
   retry: 'Retry',
   sign_in: 'SIGN IN',
@@ -51,7 +33,5 @@ export default {
   something_went_wrong: 'Something went wrong',
   submit: 'Submit',
   try_agin: 'Something went wrong. Please try again.',
-  verification: 'VERIFICATION',
-  view_more: 'VIEW MORE',
   whats_new: "What's new",
 };
