+++ package.json
@@ -14,15 +14,15 @@
     "hoist-non-react-statics": "^2.3.1",
     "lodash": "^4.17.5",
     "moment": "^2.21.0",
-    "npm": "^5.8.0",
+    "npm": "^5.7.1",
     "prop-types": "^15.6.1",
-    "react": "16.3.1",
-    "react-native": "0.55.2",
+    "react": "16.2.0",
+    "react-native": "0.53.0",
     "react-native-fabric": "^0.5.1",
     "react-native-google-analytics-bridge": "^5.6.0",
     "react-native-i18n": "^2.0.12",
     "react-native-timer": "^1.3.1",
-    "react-native-vector-icons": "^4.6.0",
+    "react-native-vector-icons": "^4.4.2",
     "react-navigation": "^1.0.0",
     "react-navigation-redux-helpers": "^1.0.1",
     "react-redux": "^5.0.6",
@@ -50,8 +50,6 @@
     "preset": "react-native"
   },
   "rnpm": {
-    "assets": [
-      "./assets/fonts/"
-    ]
+    "assets": ["./assets/fonts/"]
   }
 }
